Optimize cpu usage when task run a long time
Previously the cpu usage will be high if a task takes a long time to
run. That is because the running task will always within the work quese
and cannot be locked by other threads. So these threads will continue to
try to lock and fail over again. It wait a lot of cpu to do the trying.

Signed-off-by: Tao Liu <ltao@redhat.com>